Added the three primary html files
index.html acts as a dispatcher for the other two files.
made.html is displayed when a minified url has been made,
this will show the link and allow for easy copying.
main.html is displayed in all other circumstances and will
provide an interface for making new short urls as well as
display error messages sent back by the server.