Added the default css files
These css files should be overwritten by the user with any
branding specific styles or other design decisions.
They are only here so the service can be shipped as a complete
product.
app.css has all of the app specific styling, whatever is the same
across all of the pages.
made.css has all of the made page specific styling such as word
wrapping for longer urls.
main.css has all of the main page specific styling such as styles
for the submit button and error message.
In the future error message may be defaulted to a note pop down.